Crate problem056

Problem 56 (Powerful digit sum)

Solution detail

The only thing worth noting here is that the result of the exponentiation will never fit into a primitive type, so we should use BigUint. The rest is just a matter of doing the calculations and summing the digits, keeping track of the best digital sum as we go.

Functions

answer
Solve the problem, returning the answer as a String

main

solve
Find the largest digital sum among numbers of the form a^b, with a, b less than the given limit.

sum_of_digits
Calculate the sum of the digits (in base 10) of the given number.